requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
requests==2.29.0
scikit-learn>=1.2.2
numpy>=1.24.2
pandas>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles==23.1.0
httpx>=0.24.0
pydantic==1.10.7
nltk>=3.8.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
google-cloud-language>=2.11.0 